backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-multipart
PyMySQL
email-validator
pytest
